Fix calculation of overlap_simple_bit_stream_size
This fixes the "Stream truncated" error seen with a number of woff2
fonts. Currently the calculation of overlap_simple_bit_stream_size uses
the same function as for bitmap_stream_length but in the specification
the latter should be rounded up to the next 32bit boundary, and the
former should be only to the next byte boundary.

This simplifies/removes the bit_stream_byte_length function and uses the
same inline calculation for this as the the woff2 reference decoder.